+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.4] - 2022-02-06
+### Added
+- Numbers.isIntegral on number type also checks for primitives number types.
+- Numbers.isReal on number type also checks for primitives number types.
+
 ## [0.3.3] - 2021-04-30
 ### Added
 - Convert.toType can convert numbers to dates.